test: move debugger test case to parallel
Move test case that does not require a predetermined port to parallel.

PR-URL: https://github.com/nodejs/node/pull/39300
